consumer-python/requirements.txt
kafka
python-dotenv
lz4
